agents/pb-gh-mcp-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-agent"
version = "0.1.0"
description = "Azure AI Foundry MCP tool agent"
requires-python = ">=3.10"
dependencies = [
  "azure-ai-projects==2.0.0b3",
  "azure-ai-evaluation>=1.0.0b1",
  "azure-core-tracing-opentelemetry>=1.0.0b12",
  "azure-identity>=1.15.0",
  "azure-monitor-opentelemetry>=1.4.0",
  "openai>=1.0.0",
  "azure-core>=1.29.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "python-dotenv>=1.0.0",
  "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: integration tests requiring Azure credentials",
  "unit: fast unit tests",
]
filterwarnings = [
  "ignore:You should use `Logger` instead\\.:DeprecationWarning",
  "ignore:You should use `LoggerProvider` instead\\.:DeprecationWarning",
  "ignore:You should use `ProxyLoggerProvider` instead\\.:DeprecationWarning",
]
